requirements.txt
Flask
requests
PyMuPDF
azure-ai-textanalytics
fuzzywuzzy
python-dotenv
gunicorn
python-Levenshtein
flask-cors
python-dotenv